+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - fixed add multiple features in the live manager
 - experiences suggestions
 - terrifying can increase difficult
+- compendium-stats can find correct features now
 
 # 0.1.5
 - small css fix